arm32ve : change static function that is not called from outside
